+++ examples/apollo-watch-fragments/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - apollo-watch-fragments
 
-This log was last generated on Tue, 30 Jan 2024 10:24:57 GMT and should not be manually modified.
+This log was last generated on Wed, 28 Feb 2024 17:44:01 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.1.0
+
+Wed, 28 Feb 2024 17:44:01 GMT
+
+### Patches
+
+- Bump @graphitation/embedded-document-artefact-loader to v0.7.0
+
+## 0.1.0
+
+Tue, 06 Feb 2024 17:55:58 GMT
+
+### Patches
+
+- Bump @graphitation/apollo-react-relay-duct-tape to v1.0.30
 
 ## 0.1.0